Conflicts:
	Engine/Binaries/DotNET/GitDependencies.exe
	Engine/Binaries/DotNET/GitDependencies.exe.config
	Engine/Build/BatchFiles/Linux/Setup.sh
	Engine/Build/BatchFiles/Mac/GitDependencies.sh
	Engine/Build/BatchFiles/Mac/GitDependenciesHook.sh
	Engine/Build/BatchFiles/Mac/SetupMono.sh
	Engine/Build/Commit.gitdeps.xml
	Engine/Build/IOS/UE4Game-Info.plist
	Engine/Build/Promoted.gitdeps.xml
	Engine/Config/BaseEngine.ini
	Engine/Plugins/Media/AvfMedia/Source/AvfMedia/Private/Player/AvfMediaPlayer.cpp
	Engine/Shaders/BasePassForForwardShadingPixelShader.usf
	Engine/Shaders/SimpleElementPixelShader.usf
	Engine/Source/Developer/iOS/IOSPlatformEditor/Private/IOSTargetSettingsCustomization.cpp
	Engine/Source/Developer/iOS/IOSPlatformEditor/Private/IOSTargetSettingsCustomization.h
	Engine/Source/Editor/BlueprintGraph/Private/BlueprintActionFilter.cpp
	Engine/Source/Editor/BlueprintGraph/Private/K2Node_MCDelegate.cpp
	Engine/Source/Editor/FontEditor/Private/SCompositeFontEditor.cpp
	Engine/Source/Editor/LandscapeEditor/Private/LandscapeEdModeComponentTools.cpp
	Engine/Source/Editor/Persona/Private/SAnimMontagePanel.h
	Engine/Source/Editor/Persona/Private/SAnimNotifyPanel.cpp
	Engine/Source/Editor/UMGEditor/Private/Customizations/DetailCustomizations.cpp
	Engine/Source/Editor/UMGEditor/Private/Hierarchy/SHierarchyViewItem.cpp
	Engine/Source/Editor/UMGEditor/Public/WidgetBlueprint.h
	Engine/Source/Editor/UnrealEd/Classes/Editor/EditorEngine.h
	Engine/Source/Editor/UnrealEd/Classes/Settings/EditorExperimentalSettings.h
	Engine/Source/Editor/UnrealEd/Private/CookOnTheFlyServer.cpp
	Engine/Source/Editor/UnrealEd/Private/Dialogs/SMeshMergingDialog.cpp
	Engine/Source/Editor/UnrealEd/Private/EditorAnimUtils.cpp
	Engine/Source/Editor/UnrealEd/Private/Fbx/FbxMainImport.cpp
	Engine/Source/Editor/UnrealEd/Private/Kismet2/Kismet2.cpp
	Engine/Source/Editor/UnrealEd/Private/UnrealEdEngine.cpp
	Engine/Source/Editor/UnrealEd/Public/Dialogs/SMeshMergingDialog.h
	Engine/Source/Editor/UnrealEd/UnrealEd.Build.cs
	Engine/Source/Programs/AutomationTool/Android/AndroidPlatform.Automation.cs
	Engine/Source/Programs/AutomationTool/Scripts/BuildCommonTools.Automation.cs
	Engine/Source/Programs/AutomationTool/Scripts/GUBP.Automation.cs
	Engine/Source/Programs/GitDependencies/Program.cs
	Engine/Source/Programs/IOS/iPhonePackager/CodeSigning.cs
	Engine/Source/Programs/IOS/iPhonePackager/iPhonePackager.cs
	Engine/Source/Programs/UnrealBuildTool/Android/UEBuildAndroid.cs
	Engine/Source/Programs/UnrealBuildTool/Android/UEDeployAndroid.cs
	Engine/Source/Programs/UnrealBuildTool/Configuration/UEBuildTarget.cs
	Engine/Source/Programs/UnrealBuildTool/HTML5/HTML5SDKInfo.cs
	Engine/Source/Programs/UnrealDocTool/APIDocTool/APIDocTool/APIVariable.cs
	Engine/Source/Programs/UnrealDocTool/APIDocTool/APIDocTool/Program.cs
	Engine/Source/Programs/UnrealHeaderTool/Private/HeaderParser.cpp
	Engine/Source/Programs/UnrealVersionSelector/Private/UnrealVersionSelector.cpp
	Engine/Source/Runtime/AIModule/Private/EnvironmentQuery/Generators/EnvQueryGenerator_BlueprintBase.cpp
	Engine/Source/Runtime/AIModule/Private/Perception/AIPerceptionSystem.cpp
	Engine/Source/Runtime/AIModule/Private/Perception/AISense_Hearing.cpp
	Engine/Source/Runtime/Core/Private/Apple/ApplePlatformSymbolication.cpp
	Engine/Source/Runtime/Core/Private/Linux/LinuxApplication.cpp
	Engine/Source/Runtime/Core/Private/Linux/LinuxCursor.cpp
	Engine/Source/Runtime/Core/Private/Linux/LinuxPlatformFile.cpp
	Engine/Source/Runtime/Core/Private/Linux/LinuxPlatformSplash.cpp
	Engine/Source/Runtime/Core/Private/Windows/WindowsTextInputMethodSystem.cpp
	Engine/Source/Runtime/Core/Public/Math/Plane.h
	Engine/Source/Runtime/CoreUObject/Private/UObject/Obj.cpp
	Engine/Source/Runtime/CoreUObject/Private/UObject/PropertyBaseObject.cpp
	Engine/Source/Runtime/Engine/Classes/Components/ActorComponent.h
	Engine/Source/Runtime/Engine/Classes/Engine/World.h
	Engine/Source/Runtime/Engine/Classes/GameFramework/GameNetworkManager.h
	Engine/Source/Runtime/Engine/Private/AI/Navigation/RecastNavMeshGenerator.cpp
	Engine/Source/Runtime/Engine/Private/Animation/AnimMontage.cpp
	Engine/Source/Runtime/Engine/Private/Animation/AnimSequence.cpp
	Engine/Source/Runtime/Engine/Private/GameEngine.cpp
	Engine/Source/Runtime/Engine/Private/GameViewportClient.cpp
	Engine/Source/Runtime/Engine/Private/MovementComponent.cpp
	Engine/Source/Runtime/Engine/Private/Slate/SceneViewport.cpp
	Engine/Source/Runtime/Engine/Private/UnrealEngine.cpp
	Engine/Source/Runtime/Engine/Private/UserInterface/Console.cpp
	Engine/Source/Runtime/Engine/Private/UserInterface/EngineFontServices.cpp
	Engine/Source/Runtime/Engine/Private/World.cpp
	Engine/Source/Runtime/Engine/Public/AI/Navigation/PImplRecastNavMesh.h
	Engine/Source/Runtime/Engine/Public/AI/Navigation/RecastNavMeshGenerator.h
	Engine/Source/Runtime/Engine/Public/CanvasTypes.h
	Engine/Source/Runtime/Engine/Public/EngineFontServices.h
	Engine/Source/Runtime/Engine/Public/VisualLogger/VisualLogger.inl
	Engine/Source/Runtime/IOS/IOSRuntimeSettings/Classes/IOSRuntimeSettings.h
	Engine/Source/Runtime/Landscape/Private/Landscape.cpp
	Engine/Source/Runtime/Launch/Private/Android/LaunchAndroid.cpp
	Engine/Source/Runtime/Launch/Resources/Version.h
	Engine/Source/Runtime/Online/Android/OnlineSubsystemGooglePlay/Private/OnlineSubsystemGooglePlay.cpp
	Engine/Source/Runtime/OpenGLDrv/Private/Mac/OpenGLMac.cpp
	Engine/Source/Runtime/OpenGLDrv/Private/Mac/OpenGLMac.h
	Engine/Source/Runtime/OpenGLDrv/Public/OpenGL3.h
	Engine/Source/Runtime/Renderer/Private/PostProcess/SceneRenderTargets.cpp
	Engine/Source/Runtime/Renderer/Private/PostProcess/SceneRenderTargets.h
	Engine/Source/Runtime/Renderer/Private/SceneCaptureRendering.cpp
	Engine/Source/Runtime/Renderer/Private/ScenePrivate.h
	Engine/Source/Runtime/Slate/Private/Framework/Text/TextLayout.cpp
	Engine/Source/Runtime/Slate/Public/Widgets/Input/SCheckBox.h
	Engine/Source/Runtime/SlateCore/Private/Textures/TextureAtlas.cpp
	Engine/Source/Runtime/SlateCore/Public/Textures/TextureAtlas.h
	Engine/Source/Runtime/SlateRHIRenderer/Private/SlateRHIRenderer.cpp
	Engine/Source/Runtime/UMG/Private/Components/TextBlock.cpp
	Engine/Source/Runtime/UMG/Private/Components/WidgetComponent.cpp
	Engine/Source/Runtime/UMG/Public/Blueprint/WidgetBlueprintLibrary.h
	Engine/Source/Runtime/UMG/Public/Components/TextBlock.h
	Engine/Source/ThirdParty/WebSockets/WebSockets.Build.cs
	README.md
	Setup.command
	Setup.sh

+++ Engine/Source/Programs/AutomationTool/HTML5/HTML5Platform.Automation.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 1998-2014 Epic Games, Inc. All Rights Reserved.
+﻿// Copyright 1998-2015 Epic Games, Inc. All Rights Reserved.
 using System;
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
@@ -44,7 +44,17 @@ public class HTML5Platform : Platform
             {
                 GameExe += "-HTML5-" + Params.ClientConfigsToBuild[0].ToString();
             }
-            GameExe += ".js";
+			GameExe += ".js";
+			
+			// ensure the ue4game binary exists, if applicable
+			string FullGameExePath = Path.Combine(Path.GetDirectoryName(Params.ProjectGameExeFilename), GameExe);
+			if (!SC.IsCodeBasedProject && !FileExists_NoExceptions(FullGameExePath))
+			{
+				Log("Failed to find game application " + FullGameExePath);
+				AutomationTool.ErrorReporter.Error("Stage Failed.", (int)AutomationTool.ErrorCodes.Error_MissingExecutable);
+				throw new AutomationException("Could not find application {0}. You may need to build the UE4 project with your target configuration and platform.", FullGameExePath);
+			}
+
             if (Path.Combine(Path.GetDirectoryName(Params.ProjectGameExeFilename), GameExe) != Path.Combine(PackagePath, GameExe))
             {
                 File.Copy(Path.Combine(Path.GetDirectoryName(Params.ProjectGameExeFilename), GameExe), Path.Combine(PackagePath, GameExe), true);
@@ -83,8 +93,8 @@ public class HTML5Platform : Platform
             string OutDir = PackagePath;
             File.Copy(JSDir + "/json2.js", OutDir + "/json2.js", true);
             File.SetAttributes(OutDir + "/json2.js", FileAttributes.Normal);
-            File.Copy(JSDir + "/jstorage.js", OutDir + "/jstorage.js", true);
-            File.SetAttributes(OutDir + "/jstorage.js", FileAttributes.Normal);
+            File.Copy(JSDir + "/jStorage.js", OutDir + "/jStorage.js", true);
+            File.SetAttributes(OutDir + "/jStorage.js", FileAttributes.Normal);
             File.Copy(JSDir + "/moz_binarystring.js", OutDir + "/moz_binarystring.js", true);
             File.SetAttributes(OutDir + "/moz_binarystring.js", FileAttributes.Normal);
             PrintRunTime();
@@ -127,7 +137,7 @@ public class HTML5Platform : Platform
 				{
 					InArguments = InArguments.Replace ("\"", "");
 					string[] Arguments = InArguments.Split(' ');
-					string ArgumentString = IsContentOnly ? "'" + InGameName + "/" + InGameName + ".uproject '," : "";
+					string ArgumentString = IsContentOnly ? "'../../../" + InGameName + "/" + InGameName + ".uproject '," : "";
 					for (int i = 0; i < Arguments.Length - 1; ++i)
 					{
 						ArgumentString += "'";
@@ -194,7 +204,7 @@ public class HTML5Platform : Platform
 		SC.ArchiveFiles(PackagePath, Path.GetFileName(GameExe));
 		SC.ArchiveFiles(PackagePath, Path.GetFileName(GameExe + ".mem"));
 		SC.ArchiveFiles(PackagePath, Path.GetFileName("json2.js"));
-		SC.ArchiveFiles(PackagePath, Path.GetFileName("jstorage.js"));
+		SC.ArchiveFiles(PackagePath, Path.GetFileName("jStorage.js"));
 		SC.ArchiveFiles(PackagePath, Path.GetFileName("moz_binarystring.js"));
 		SC.ArchiveFiles(PackagePath, Path.GetFileName(OutputFile));
 	}
@@ -285,10 +295,9 @@ public class HTML5Platform : Platform
 			ClientProcess.ProcessObject.Exited += delegate(System.Object o, System.EventArgs e)
 			{
 				System.Console.WriteLine("Browser Process Ended (PID={0})", ClientProcess.ProcessObject.Id);
-				var bFoundChildProcess = true;
+				var bFoundChildProcess = false;
 				if (bBrowserWillSpawnProcess)
 				{
-					bFoundChildProcess = false;
 					// Chrome spawns a process from the tab it opens and then lets the process we spawned die, so
 					// catch that process and attach to that instead.
 					var CurrentProcesses = Process.GetProcesses();
@@ -313,7 +322,7 @@ public class HTML5Platform : Platform
 
 				if (!bFoundChildProcess)
 				{
-					System.Console.WriteLine("- Killing Webserver", ClientProcess.ProcessObject.Id);
+					System.Console.WriteLine("- Killing Webserver PID({0})", Result.ProcessObject.Id);
 					Result.ProcessObject.StandardInput.Close();
 					Result.ProcessObject.Kill();
 				}
@@ -358,8 +367,13 @@ public class HTML5Platform : Platform
 
     public override PakType RequiresPak(ProjectParams Params)
     {
-        return PakType.Never;
+        return PakType.Always;
     }
+
+	public override string GetPlatformPakCommandLine()
+	{
+		return " -compress";
+	}
 
 	public override bool DeployLowerCaseFilenames(bool bUFSFile)
 	{
